feat: complete part 2
Added button to start and restart game.

Added a property to the Game class for keeping track of when the game is
over.

Most of the changes for handleClick(evt) is from indentation.  Only one
line was added in the beginning, which was a check for when the game is
over.