feat: improve JavaScript
Changed the Game class constructor to take arguments for players.  This
should make the class more flexible.

For the Game class, placed players into an array instead of their own
variables.  This will reduce the amount of characters that need to be
written and have all players stored in one single place.

Created a bound method handleGameClick so that the event listener for
the top table row that's used for dropping in pieces can more easily be
removed.

Removed the event listener, for handling clicks on the top row of the
table, after the game is over.  This should have better performance
than continually checking for the game over state every time the top
table row is clicked.

For handleClick(evt), only the first line, with the check for game
over, was removed.  The rest is indentation changes made by Prettier.